AdaptiveSVC: Implement Adaptive SVC
This algorithm selects the best classifier among SVC trained on different window
sizes. A window is a number of stream batch. Training on a window size 1 means
training on the last batch of data received. The prediction is then made by the
selected classifier, alone.

The expected general behaviour under concept drift is :
- To increase the window size when the concept is stable
- To decrease the window size when the concept is changing
(The more sudden is the concept drift, the quicker is the decrease)